peer: move connection wrapper in its own file
