+++ pkg/common/track_info.go
@@ -75,21 +75,3 @@ func TrackInfoFromTrack(track *webrtc.TrackRemote) TrackInfo {
 		Layer:    RIDToSimulcastLayer(track.RID()),
 	}
 }
-
-type ConnectionWrapper struct {
-	connection *webrtc.PeerConnection
-}
-
-func NewConnectionWrapper(connection *webrtc.PeerConnection) ConnectionWrapper {
-	return ConnectionWrapper{
-		connection: connection,
-	}
-}
-
-func (c ConnectionWrapper) Subscribe(track *webrtc.TrackLocalStaticRTP) (*webrtc.RTPSender, error) {
-	return c.connection.AddTrack(track)
-}
-
-func (c ConnectionWrapper) Unsubscribe(sender *webrtc.RTPSender) error {
-	return c.connection.RemoveTrack(sender)
-}
